Supply missing double/rational comparisons
The original code provided half of the comparison operators for
cross-type, RationalUDT vs. double comparisons: it provided the
half where the rational is on the left. Although the operators
named their commutators, said commutators weren't provided, which
would lead to errors at runtime on attempts to use them.

The change to synthetic operator notation changed which operators
were missing, but still supplied only half. From < (rational,double)
was synthesized its commutator, > (double,rational), but still there
was neither a < (double,rational) nor a > (rational,double). And so
on down the line.

For a full set of cross-type comparisons on types A,B, it is still
necessary to supply four implementing methods: one of < <= >= > and
one of = <>, on each of A,B and B,A. The remaining eight comparison
operators can be synthesized from those to complete the set.