Make operators more DRY with TWIN notation
In addition to commutator=SELF where an operator (with both operands
the same type) is its own commutator, commutator=TWIN applies where
an operator is cross-type, and its corresponding types-commuted
operator should have the same name (which is allowed, thanks to
the exchanged operand types). Both notations reduce the risk of
copy-pastos by reducing the need to repeat the operator name.

TWIN also can be used as the value of synthetic= to indicate that
the new synthetic function should have the same name as the function
it is based on. This is only possible when the two operand types
are not the same, so the commuted version will have a different
signature, and of course only makes sense for a function that's
commutative. Naturally, SELF is not a valid value for synthetic=.